+++ src/main/java/com/darigan/lil/learningspring/data/GuestRepository.java
@@ -2,5 +2,5 @@ package com.darigan.lil.learningspring.data;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface GuestRepository extends CrudRepository<Guest, Long> {
+public interface GuestRepository extends CrudRepository<Guest, Integer> {
 }
